get gameDetail from steamApi

+++ src/main/java/com/nextgamezada/steamApp/SteamApp.java
@@ -2,20 +2,20 @@ package com.nextgamezada.steamApp;
 
 public class SteamApp {
 
-    private int appid;
+    private long appid;
 
     private String name;
 
-    public SteamApp(int appid, String name) {
+    public SteamApp(long appid, String name) {
         this.appid = appid;
         this.name = name;
     }
 
-    public int getAppid() {
+    public long getAppid() {
         return appid;
     }
 
-    public void setAppid(int appid) {
+    public void setAppid(long appid) {
         this.appid = appid;
     }
